Time: 0 ms (100.00%) | Memory: 40.8 MB (29.79%) - LeetSync

+++ 2649-count-total-number-of-colored-cells/count-total-number-of-colored-cells.java
@@ -1,6 +1,6 @@
 class Solution {
     
     public long coloredCells(int n){
-        return (long)1+(long)2*(long)n*(long)(n-1);
+        return 1+2*(long)n*(long)(n-1);
     }
 }
